Make DJVU the default
Firstly DJVU is often easier to work with at Wikisource than PDF,
the thumbnails render much faster and the text layer is generally
better.

Also a massive PDF upload means it's highly likely that a user is
looking to upload a DVJU in preference to the PDF (if they wanted
a PDF, that's probably already at Commons).